SSCS-7578 Moving esa specific fields into SscsEsaCaseData and generic fields up to SscsCaseData
