+++ src/main/java/uk/gov/hmcts/reform/sscs/ccd/domain/SscsCaseData.java
@@ -6,7 +6,6 @@ import com.fasterxml.jackson.annotation.*;
 import com.fasterxml.jackson.databind.annotation.JsonSerialize;
 import com.fasterxml.jackson.datatype.jsr310.ser.LocalDateSerializer;
 import java.time.LocalDate;
-import java.util.ArrayList;
 import java.util.Arrays;
 import java.util.Collection;
 import java.util.Collections;
@@ -308,11 +307,9 @@ public class SscsCaseData implements CaseData {
     @JsonProperty("supportGroupOnlyAppeal")
     private String supportGroupOnlyAppeal;
     @JsonUnwrapped
+    @Getter(AccessLevel.NONE)
     private SscsEsaCaseData sscsEsaCaseData;
-    private YesNo doesRegulation29Apply;
-    private YesNo showRegulation29Page;
-    private YesNo showSchedule3ActivitiesPage;
-    private YesNo doesRegulation35Apply;
+    private String dwpReassessTheAward;
     private YesNo showFinalDecisionNoticeSummaryOfOutcomePage;
 
     @JsonIgnore
@@ -480,34 +477,6 @@ public class SscsCaseData implements CaseData {
             this.translationWorkOutstanding = "No";
         } else {
             this.translationWorkOutstanding = "Yes";
-        }
-    }
-
-    @JsonIgnore
-    public YesNo getRegulation35Selection() {
-        if (sscsEsaCaseData != null) {
-            if ("No".equalsIgnoreCase(sscsEsaCaseData.getEsaWriteFinalDecisionSchedule3ActivitiesApply())) {
-                return doesRegulation35Apply;
-            } else {
-                return null;
-            }
-        } else {
-            return null;
-        }
-    }
-
-    @JsonIgnore
-    public List<String> getSchedule3Selections() {
-        if (sscsEsaCaseData != null) {
-            if ("Yes".equalsIgnoreCase(sscsEsaCaseData.getEsaWriteFinalDecisionSchedule3ActivitiesApply())) {
-                return sscsEsaCaseData.getEsaWriteFinalDecisionSchedule3ActivitiesQuestion();
-            } else if ("No".equalsIgnoreCase(sscsEsaCaseData.getEsaWriteFinalDecisionSchedule3ActivitiesApply())) {
-                return new ArrayList<>();
-            } else {
-                return null;
-            }
-        } else {
-            return null;
         }
     }
