+++ src/test/java/uk/gov/hmcts/reform/sscs/ccd/service/SscsCcdConvertServiceTest.java
@@ -277,8 +277,8 @@ public class SscsCcdConvertServiceTest {
                 .build();
         SscsCaseDetails caseDetails = new SscsCcdConvertService().getCaseDetails(build);
 
-        assertEquals(YesNo.YES, caseDetails.getData().getShowRegulation29Page());
-        assertEquals(YesNo.NO, caseDetails.getData().getDoesRegulation29Apply());
+        assertEquals(YesNo.YES, caseDetails.getData().getSscsEsaCaseData().getShowRegulation29Page());
+        assertEquals(YesNo.NO, caseDetails.getData().getSscsEsaCaseData().getDoesRegulation29Apply());
         assertTrue(YesNo.YES.toBoolean());
         assertFalse(YesNo.NO.toBoolean());
     }
